broke out app for easier testing
